pyproject.toml
[project]
name = "resummarized-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crispy-daisyui>=0.8.0",
    "django-components>=0.143.2",
    "django-crispy-forms>=2.5",
    "django-markdownify>=0.9.5",
    "django-tailwind[cookiecutter,honcho,reload]>=4.4.1",
    "dotenv>=0.9.9",
    "google-genai>=1.50.1",
    "heroicons>=2.13.0",
    "pgvector>=0.4.1",
    "psycopg2>=2.9.11",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
    "transformers>=4.57.1",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu" },
]
torchvision = [
  { index = "pytorch-cpu" },
]
